fix unittest structure
