+++ package/tests/test_cp/test_azure/test_common/test_validtors/test_validator_factory.py
@@ -17,7 +17,4 @@ class TestValidators(TestCase):
         self.assertRaises(Exception,
                           validator_factory.try_validate,
                           StorageAccount,
-                          None
-                          )
-
-
+                          None)
